Rename be.guanaco.smx.scheduler -> io.guanaco.scheduler
