--- api/src/main/java/be/guanaco/smx/scheduler/TaskStatus.java
@@ -1,10 +1,0 @@
-package be.guanaco.smx.scheduler;
-
-/**
- * Enumeration of possible task status values.
- */
-public enum TaskStatus {
-
-    IDLE, ACTIVE;
-
-}
