+++ api/src/main/java/io/guanaco/scheduler/TaskStatus.java
@@ -1,0 +1,10 @@
+package io.guanaco.scheduler;
+
+/**
+ * Enumeration of possible task status values.
+ */
+public enum TaskStatus {
+
+    IDLE, ACTIVE;
+
+}
