+++ camel/src/main/java/io/guanaco/scheduler/camel/JavaCamelTaskRouteBuilder.java
@@ -1,0 +1,32 @@
+package io.guanaco.scheduler.camel;
+
+import io.guanaco.alerta.api.Alerta;
+import io.guanaco.scheduler.Task;
+import org.apache.camel.builder.RouteBuilder;
+import scala.collection.JavaConversions;
+import scala.collection.Seq;
+
+import java.util.List;
+
+/**
+ * Abstract Camel {@link RouteBuilder} implementation to implement an {@link Task} using Camel routes.
+ * It also provides out-of-the-box integration with Alerta (sending heartbeats for every run of the task).
+ */
+public abstract class JavaCamelTaskRouteBuilder extends CamelTaskRouteBuilder {
+
+
+    public JavaCamelTaskRouteBuilder(String identifier, String schedule, Alerta alerta) {
+        super(identifier, schedule, alerta);
+    }
+
+    @Override
+    public Seq<CamelTaskItem> camelTaskItems() {
+        return JavaConversions.asScalaBuffer(getCamelTaskItems()).toSeq();
+    }
+
+    /**
+     * Get the task items that need to be handled by this scheduled task.
+     */
+    protected abstract List<CamelTaskItem> getCamelTaskItems();
+
+}
