+++ pom.xml
@@ -4,7 +4,7 @@
 
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>be.guanaco.smx</groupId>
+    <groupId>io.guanaco</groupId>
     <artifactId>scheduler</artifactId>
     <version>1.0.1-SNAPSHOT</version>
 
@@ -42,7 +42,7 @@
         <dependencies>
             <!-- Project dependencies -->
             <dependency>
-                <groupId>be.guanaco.smx.scheduler</groupId>
+                <groupId>io.guanaco.scheduler</groupId>
                 <artifactId>api</artifactId>
                 <version>${project.version}</version>
             </dependency>
