+++ rest/pom.xml
@@ -5,12 +5,12 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>be.guanaco.smx</groupId>
+        <groupId>io.guanaco</groupId>
         <artifactId>scheduler</artifactId>
         <version>1.0.1-SNAPSHOT</version>
     </parent>
 
-    <groupId>be.guanaco.smx.scheduler</groupId>
+    <groupId>io.guanaco.scheduler</groupId>
     <artifactId>rest</artifactId>
 
     <packaging>bundle</packaging>
@@ -18,7 +18,7 @@
 
     <dependencies>
         <dependency>
-            <groupId>be.guanaco.smx.scheduler</groupId>
+            <groupId>io.guanaco.scheduler</groupId>
             <artifactId>api</artifactId>
         </dependency>
         <dependency>
@@ -42,7 +42,7 @@
                 <artifactId>maven-bundle-plugin</artifactId>
                 <configuration>
                     <instructions>
-                        <Private-Package>be.guanaco.smx.scheduler.rest</Private-Package>
+                        <Private-Package>io.guanaco.scheduler.rest</Private-Package>
                     </instructions>
                 </configuration>
             </plugin>
